refactor(ref_parser): move module to mod.rs with dir support
Refactor the ref_parser module by relocating all its contents from src/ref_parser.rs into
src/ref_parser/mod.rs, turning it into a submodule. Added optional directory parameters to
internal implementations for functions like current_branch_impl, repo_from_fetch_head_impl, and repo_from_origin_impl
allowing git commands to run in specific directories. This supports more flexible and testable
code. The original src/ref_parser.rs file and its corresponding tests.rs file are removed
and their contents moved to new locations with continued coverage and test fixtures.
